requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
httplib2==0.22.0
idna==3.7
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
py-cord==2.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
